pyproject.toml
[project]
name = "bg-qqbot"
version = "0.1.0"
description = "炉石战棋联赛 QQ 机器人"
requires-python = ">=3.11"
dependencies = [
    "nonebot2[fastapi]>=2.3.0",
    "nonebot-adapter-onebot>=2.4.0",
    "httpx>=0.27.0",
    "nonebot-plugin-apscheduler>=0.4.0",
]

[tool.nonebot]
adapters = [
    { name = "OneBot V11", module_name = "nonebot.adapters.onebot.v11" }
]
plugins = []
plugin_dirs = ["src/plugins"]
